onboarding/requirements.txt
streamlit==1.29.0
pandas==2.1.4
google-cloud-bigquery==3.13.0
google-cloud-secret-manager==2.17.0
google-api-python-client==2.111.0
google-cloud-run==0.10.0
pandas-gbq==0.19.2
python-dotenv==1.0.0
pytz==2023.3.post1
python-dateutil==2.8.2